Extending Features Sheets
Added dedicated Feature Sheet to both Icewind Dale Sheet and the Mythos Sheet. In order to not waste the space on the Mythos sheet that is about half a page for features I added a new field "Features-and-traits-3" for the full pages on Mythos and kept the normal "features-and-traits-2" but limited it to 500 char to avoid it shrinking to unreadable sizes.